projects/converter/src/requirements.txt
flask==2.2.2
pymongo==4.3.3
requests==2.28.1
pydantic==1.10.4
pika==1.3.1
PyPDF2==3.0.1
gTTS==2.3.1
PyJWT==2.6.0